Minor changes
Small chagnes in CMakeLists. Also worked a little bit on warnings.

+++ Omniforce/src/Core/Application.h
@@ -58,9 +58,9 @@ namespace Omni {
 		ImGuiRenderer* m_ImGuiRenderer;
 
 		struct DeltaTimeData {
-			float64 delta_time = 1.0f / 60.0f;
-			float64 last_frame_time = 0.0f;
-			float64 current_frame_time = 60.0f;
+			float32 delta_time = 1.0f / 60.0f;
+			float32 last_frame_time = 0.0f;
+			float32 current_frame_time = 60.0f;
 		} m_DeltaTimeData;
 	};
